backend/ai-service/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
mongoengine==0.27.0
python-dotenv==1.0.0
PyJWT==2.8.0
pymongo==4.3.3
requests==2.31.0
google-generativeai==0.3.2
Pillow==10.2.0